% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framework.R
\name{method_name}
\alias{method_name}
\title{Label a framework method}
\usage{
method_name(u, l1, g_cph, se_cph)
}
\arguments{
\item{u, l1}{Framework parameters.}

\item{g_cph, se_cph}{Historical effect estimate and standard error.}
}
\value{
A character label.
}
\description{
Label a framework method
}
