% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nidesigns-package.R
\docType{package}
\name{nidesigns-package}
\alias{nidesigns}
\alias{nidesigns-package}
\title{nidesigns: Active-Controlled Non-Inferiority Trial Design}
\description{
Implements the general framework described by Olivas-Martinez, Gao, and
Janes (2026). The historical and active-controlled trial effect estimators
are assumed independent and approximately normal on an additive analysis
scale, with smaller values indicating benefit.
}
\references{
Olivas-Martinez A, Gao F, Janes H (2026). "A General Framework for
Designing and Evaluating Active-Controlled Trials with Non-Inferiority
Objectives." \emph{Statistics in Medicine}, 45, e70618.
\doi{10.1002/sim.70618}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aolivasm/ni-designs}
  \item Report bugs at \url{https://github.com/aolivasm/ni-designs/issues}
}

}
\author{
\strong{Maintainer}: Antonio Olivas-Martinez \email{aolivas@uw.edu}

Other contributors:
\itemize{
  \item Fei Gao [contributor]
  \item Holly Janes [contributor]
}

}
\keyword{internal}
